+++ src/components/Card/MakeMyCard/MakeMyCard.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MakeMyCard = () => {
+  return <div>MakeMyCard</div>;
+};
+
+export default MakeMyCard;
