+++ src/components/index.tsx
@@ -1,7 +1,8 @@
-export { default as Navigation } from './Navigation/Navigation';
+export { default as Header } from './Header/Header';
 export { default as Footer } from './Footer/Footer';
+export { default as Navigation } from './Navigation/Navigation';
 export { default as Button } from './Button/Button';
 export { default as IconSocial } from './Icon/IconSocial';
-export { default as Header } from './Header/Header';
-export { default as ProfileCard } from './Card/ProfileCard/ProfileCard';
 export { default as CoffeeOption } from './CoffeeOption/CoffeeOption';
+export { default as MakeMyCard } from './Card/MakeMyCard/MakeMyCard';
+export { default as ProfileCard } from './Card/ProfileCard/ProfileCard';
