+++ src/pages/index.tsx
@@ -1,3 +1,3 @@
 export { default as Home } from './Home/Home';
-export { default as MakeMyCard } from './MakeMyCard/MakeMyCard';
+export { default as MyCard } from './MyCard/MyCard';
 export { default as AllCoffes } from './AllCoffes/AllCoffes';
