+++ src/store/index.ts
@@ -1,13 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { creatorsApi } from './api/creators';
 import { contactsApi } from './api/exampleApi';
-import { makeMyCardApi } from './api/makeMyCard';
+import { myCardApi } from './api/myCard';
 
 export const store = configureStore({
   reducer: {
     [contactsApi.reducerPath]: contactsApi.reducer,
     [creatorsApi.reducerPath]: creatorsApi.reducer,
-    [makeMyCardApi.reducerPath]: makeMyCardApi.reducer,
+    [myCardApi.reducerPath]: myCardApi.reducer,
   },
   middleware: getDefaultMiddleware =>
     getDefaultMiddleware().concat(contactsApi.middleware),
